Fix ADS and DOI links in BST file
This is now working for article reference types.  But I still need to
make it work for inproceedings, book, misc, etc.  I have tested it
with rm-journal-example.tex and rmbooksamp.tex but further testing is
still needed.  Some of the functionality needs to be migrated from the
example documents to the class file.